FWGL.Graphics

Documentation

module FWGL.Graphics.Color

data Scene

data Object

data Geometry

data Texture

A texture.

data Color

An RGBA 32-bit color.

Constructors

Color !Word8 !Word8 !Word8 !Word8

data V2

Two-dimensional vector.

Constructors

V2 !Float !Float

data V3

Three-dimensional vector.

Constructors

V3 !Float !Float !Float

mkGeometry

Arguments

:: [V3]
List of vertices.

-> [V2]
List of UV coordinates.

-> [V3]
List of normals.

-> [Word16]
Triangles expressed as triples of indices to the three lists above.

-> Geometry

Creates a Geometry. The first three lists should have the same length.

mkTexture

Arguments

:: Int
Width.

-> Int
Height.

-> [Color]
List of pixels

-> Texture

Creates a Texture from a list of pixels.

textureURL

Arguments

:: String
URL

-> Texture

Creates a Texture from an URL.

vec2 :: (Float, Float) -> V2

Creates a two-dimensional vector.

vec3 :: (Float, Float, Float) -> V3

Creates a three-dimensional vector.

nothing :: Object

An empty object.

cube

Arguments

:: Float
Edge

-> Object

A cube.

geom :: Geometry -> Object

An object with a specified Geometry.

translate :: V3 -> Object -> Object

Translate an object.

rotX :: Float -> Object -> Object

Rotate an object around the X axis.

rotY :: Float -> Object -> Object

Rotate an object around the Y axis.

rotZ :: Float -> Object -> Object

Rotate an object around the Z axis.

rotAA :: V3 -> Float -> Object -> Object

Rotate an object using a rotation axis and angle.

scale :: Float -> Object -> Object

Scale an object.

scaleV :: V3 -> Object -> Object

Scale an object in three dimensions.

color :: Color -> Object -> Object

Set the color of an object.

texture :: Texture -> Object -> Object

Set the texture of an object.

dynamicE

Arguments

:: Object
Initial Object.

-> SF (Event Object) Object

Like dynamic, but instead of comparing the two objects it checks the event with the new object.

dynamic :: SF Object Object

Automatically deallocate the previous mesh from the GPU when it changes.

scene :: [Object] -> Scene

Build a scene (no projection).

perspective :: Float -> Float -> Float -> Float -> [Object] -> Scene

Build a scene (perspective).