FWGL.Graphics.Custom

Documentation

module FWGL.Vector

data Layer

An object associated with a program.

data Object gs is

An object is a set of geometries associated with some uniforms.

data AttrList is where

A set of attributes and indices.

Constructors

AttrListNil :: AttrList []

AttrListCons :: (Hashable c, AttributeCPU c g) => g -> [c] -> AttrList gs -> AttrList (g : gs)

data Geometry is

data Texture

A texture.

data Color

An RGBA 32-bit color.

Constructors

Color !Word8 !Word8 !Word8 !Word8

(~~) :: (Equal gs gs', Equal is is') => Object gs is -> Object gs' is' -> Object (Union gs gs') (Union is is')

Join two objects.

program :: (Subset gs' gs, Subset gs'' gs, Subset os' os) => VertexShader gs' is os -> FragmentShader gs'' os' -> Program gs is

Create a Program from the shaders.

nothing :: Object [] []

An empty custom object.

static :: Geometry i -> Object [] i

A custom object with a specified Geometry.

global :: (Typeable g, UniformCPU c g) => g -> c -> Object gs is -> Object (g : gs) is

Sets a global (uniform) of an object.

globalDraw :: (Typeable g, UniformCPU c g) => g -> Draw c -> Object gs is -> Object (g : gs) is

Sets a global (uniform) of an object using the Draw monad.

globalTexture :: (BackendIO, Typeable g, UniformCPU ActiveTexture g) => g -> Texture -> Object gs is -> Object (g : gs) is

Sets a global (uniform) of an object using a Texture.

globalTexSize :: (BackendIO, Typeable g, UniformCPU c g) => g -> Texture -> ((Int, Int) -> c) -> Object gs is -> Object (g : gs) is

Sets a global (uniform) of an object using the dimensions of a Texture.

layer :: (Subset oi pi, Subset og pg) => Program pg pi -> Object og oi -> Layer

Associate an object with a program.

unsafeJoin :: (Equal gs'' (Union gs gs'), Equal is'' (Union is is')) => Object gs is -> Object gs' is' -> Object gs'' is''

Join two objects, even if they don't provide the same variables.

mkGeometry :: GLES => AttrList is -> [Word16] -> Geometry is

Create a custom Geometry.

mkTexture

Arguments

:: GLES

=> Int
Width.

-> Int
Height.

-> [Color]
List of pixels

-> Texture

Creates a Texture from a list of pixels.

textureURL

Arguments

:: String
URL

-> Texture

Creates a Texture from an URL (JavaScript only).

textureFile :: String -> Texture

Creates a Texture from a file (Desktop only).

colorTex :: GLES => Color -> Texture

Generate a 1x1 texture.