FWGL.Shader.Default2D

Documentation

type Uniforms = `[View2, Image, Depth, Transform2]`

type Attributes = `[Position2, UV]`

newtype Image

An uniform that represents the texture used in the default 2D shader.

Constructors

Image Sampler2D

newtype Depth

An uniform that represents the depth used in the default 2D shader.

Constructors

Depth Float

newtype Transform2

An uniform that represents the transformation matrix used in the default 2D shader.

Constructors

Transform2 M3

newtype View2

An uniform that represents the view matrix used in the default 2D shader.

Constructors

View2 M3

newtype Position2

Constructors

Position2 V2

newtype UV

Constructors

UV V2

vertexShader :: VertexShader `[Transform2, View2, Depth]` `[Position2, UV]` `[UV]`

fragmentShader :: FragmentShader `[Image]` `[UV]`