FWGL.Vector

Documentation

data V2

Two-dimensional vector.

Constructors

V2 !Float !Float

data V3

Three-dimensional vector.

Constructors

V3 !Float !Float !Float

data V4

Constructors

V4 !Float !Float !Float !Float

data M2

Constructors

M2 !V2 !V2

data M3

Constructors

M3 !V3 !V3 !V3

data M4

Constructors

M4 !V4 !V4 !V4 !V4

vec2 :: (Float, Float) -> V2

Creates a two-dimensional vector.

vec3 :: (Float, Float, Float) -> V3

Creates a three-dimensional vector.

vec4 :: (Float, Float, Float, Float) -> V4

quat :: Float -> V3 -> V4

mat2 :: (Float, Float, Float, Float) -> M2

mat3 :: (Float, Float, Float, Float, Float, Float, Float, Float, Float) -> M3

mat4 :: (Float, Float, Float, Float, Float, Float, Float, Float, Float, Float, Float, Float, Float, Float, Float, Float) -> M4

mul4 :: M4 -> M4 -> M4

transpose4 :: M4 -> M4

idMat :: M4

transMat :: V3 -> M4

rotXMat :: Float -> M4

rotYMat :: Float -> M4

rotZMat :: Float -> M4

rotAAMat :: V3 -> Float -> M4

scaleMat :: V3 -> M4

perspectiveMat :: Float -> Float -> Float -> Float -> M4