FWGL.Geometry

Documentation

data Geometry

Constructors

Geometry

Fields

vertices :: [V3]

uvCoords :: [V2]

normals :: [V3]

elements :: [Word16]

hash :: Int

mkGeometry

Arguments

:: [V3]
List of vertices.

-> [V2]
List of UV coordinates.

-> [V3]
List of normals.

-> [Word16]
Triangles expressed as triples of indices to the three lists above.

-> Geometry

Creates a Geometry. The first three lists should have the same length.

facesToArrays :: Vector V3 -> Vector V2 -> Vector V3 -> [[(Int, Int, Int)]] -> [(V3, V2, V3)]

arraysToElements :: Foldable f => f (V3, V2, V3) -> Geometry

triangulate :: [a] -> [(a, a, a)]